create and fill postion, in_tx_order in init, use creation_date

+++ cmd/initialise/sql/cockroach/08_events_table.sql
@@ -17,7 +17,7 @@ CREATE TABLE IF NOT EXISTS eventstore.events (
 	, resource_owner TEXT NOT NULL
 	, instance_id TEXT NOT NULL
 	, "position" DECIMAL NOT NULL
-	, in_tx_order BIGINT NOT NULL
+	, in_tx_order INTEGER NOT NULL
 
 	, PRIMARY KEY (instance_id, aggregate_type, aggregate_id, event_sequence DESC)
 );
